third_party/WebKit/LayoutTests/bluetooth/service/getCharacteristic/gen-blocklisted-characteristic.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return setBluetoothFakeAdapter('BlocklistTestAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['device_information']}]}))
    .then(device => device.gatt.connect())
    .then(gatt => gatt.getPrimaryService('device_information'))
    .then(service => {
      return assert_promise_rejects_with_message(
        service.getCharacteristic('serial_number_string'),
        new DOMException('getCharacteristic(s) called with blocklisted UUID. ' +
                         'https://goo.gl/4NeimX',
                         'SecurityError'),
        'Serial Number String characteristic is blocklisted.');
    });
}, 'Serial Number String characteristic is blocklisted. Should reject with ' +
   'SecurityError.');

</script>
